more improvements to camera and rmlui

+++ src/window.cpp
@@ -92,7 +92,7 @@ namespace darmok
 		return glm::vec2(_size) / glm::vec2(_pixelSize);
 	}
 
-	glm::uvec2 WindowImpl::screenToWindowPoint(const glm::vec2& point) const noexcept
+	glm::vec2 WindowImpl::screenToWindowPoint(const glm::vec2& point) const noexcept
 	{
 		auto f = getScreenToWindowFactor();
 		auto p = (point - glm::vec2(0.5F)) * f;
@@ -100,7 +100,7 @@ namespace darmok
 		return p;
 	}
 
-	glm::vec2 WindowImpl::windowToScreenPoint(const glm::uvec2& point) const noexcept
+	glm::vec2 WindowImpl::windowToScreenPoint(const glm::vec2& point) const noexcept
 	{
 		glm::vec2 p(point);
 		p.y = _size.y - p.y;
@@ -170,12 +170,12 @@ namespace darmok
 		return _impl->getMode();
 	}
 
-	glm::uvec2 Window::windowToScreenPoint(const glm::vec2& pos) const noexcept
+	glm::vec2 Window::windowToScreenPoint(const glm::vec2& pos) const noexcept
 	{
 		return _impl->windowToScreenPoint(pos);
 	}
 
-	glm::uvec2 Window::screenToWindowPoint(const glm::vec2& pos) const noexcept
+	glm::vec2 Window::screenToWindowPoint(const glm::vec2& pos) const noexcept
 	{
 		return _impl->screenToWindowPoint(pos);
 	}
